requirements.txt
flask>=2.3.0,<3.0.0
requests>=2.31.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
pillow>=10.0.0,<11.0.0
webcolors>=1.13.0,<2.0.0
matplotlib>=3.7.0,<4.0.0